Weird hack to keep install from failing
The share install moves the entire google_appengine output that is
unzipped into the Cellar; when the install block finishes, the
install_metafiles code attempts to look for metafiles which seems
to expect that the top-level of the unpacked directory still exists.
Without this line, it tells me that the google_appengine directory
doesn't exist. This line, obviously, creates it...